Change base_vla structure
