+++ configs/gr00tn17/gr00t_n17_native_libero_10_full_finetune.py
@@ -226,7 +226,7 @@ model = dict(
     processor_kwargs=_PROCESSOR_KWARGS,
     load_metadata=True,
     qwen3_runtime='compat_457',
-    freeze_vlm_backbone=False,
+    freeze_vlm_backbone=True,
     freeze_projector=False,
     vlm_backbone=dict(
         type='GrootN17Qwen3Backbone',
